Separate gtfs rt protobuf parsing from http client code
This will make transition to mqtt easier
